Create Data.Hastory.Server.TestUtils.
